Remove a branch on the hottest path
